+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Minor Improvements
 
 * Add character count to textarea
+* Form now accepts a `onSubmit` prop which is only called when the form is valid.
+* AppWrapper now has a minium width of 958px.
 
 # 0.20.0